feat(primitive): update
