+++ config.ini
@@ -1,10 +1,22 @@
 [feature_store]
 # `feature_store.py` use this throttle to distinct `good_questions` and `bad_questions`
 reject_throttle = -1.0
-# text2vec model path, support local relative path and huggingface model format.
-# also support local path, model_path = "/path/to/your/text2vec-model"
+# text2vec model, support local relative path, huggingface repo and URL.
+# for example:
+#  "maidalun1020/bce-embedding-base_v1"
+#  "BAAI/bge-m3"
+#  "https://api.siliconflow.cn/v1/embeddings"
 embedding_model_path = "maidalun1020/bce-embedding-base_v1"
+
+# reranker model, support list:
+#  "maidalun1020/bce-reranker-base_v1"
+#  "BAAI/bge-reranker-v2-minicpm-layerwise"
+#  "https://api.siliconflow.cn/v1/rerank"
 reranker_model_path = "maidalun1020/bce-reranker-base_v1"
+
+# if `embedding_model_path` or `reranker_model_path` using `siliconcloud` API, give the token
+api_token = ""
+api_rpm = 1000
 work_dir = "workdir"
 
 [web_search]
