+++ huixiangdou/primitive/embedder.py
@@ -5,51 +5,64 @@ import pdb
 from typing import Any, List
 
 import numpy as np
-import torch
 from loguru import logger
-from .query import DistanceStrategy
+from .query import distance_strategy
+from .rpm import RPM
 
 class Embedder:
     """Wrap text2vec (multimodal) model."""
     client: Any
+    _type: str
 
-    def __init__(self, model_path: str):
+    def __init__(self, model_config: dict):
         self.support_image = False
         # bce also use euclidean distance.
         self.distance_strategy = DistanceStrategy.EUCLIDEAN_DISTANCE
         
-        if self.use_multimodal(model_path=model_path):
+        model_path = model_config['embedding_model_path']
+        self._type = self.model_type(model_path=model_path)
+        if 'bce' in self._type:
+            from sentence_transformers import SentenceTransformer
+            self.client = SentenceTransformer(model_name_or_path=model_path).half()
+        elif 'bge' in self._type:
             from FlagEmbedding.visual.modeling import Visualized_BGE
             self.support_image = True
             vision_weight_path = os.path.join(model_path, 'Visualized_m3.pth')
             self.client = Visualized_BGE(
                 model_name_bge=model_path,
                 model_weight=vision_weight_path).eval()
+        elif 'siliconcloud' in self._type:
+            self.client = {
+                'api_token': model_config['api_token'],
+                'api_rpm': model_config['api_rpm']
+            }
         else:
-            from sentence_transformers import SentenceTransformer
-            self.client = SentenceTransformer(model_name_or_path=model_path).half()
+            raise ValueError('Unknown type {}'.format(self._type))
 
     @classmethod
-    def use_multimodal(self, model_path):
+    def model_type(self, model_path):
         """Check text2vec model using multimodal or not."""
+        if model_path.startswith('https'):
+            return 'siliconcloud'
 
         if 'bge-m3' not in model_path.lower():
-            return False
+            return 'bce'
 
         vision_weight = os.path.join(model_path, 'Visualized_m3.pth')
         if not os.path.exists(vision_weight):
             logger.warning(
                 '`Visualized_m3.pth` (vision model weight) not exist')
-            return False
-        return True
+            return 'bce'
+        return 'bge'
 
     def embed_query(self, text: str = None, path: str = None) -> np.ndarray:
         """Embed input text or image as feature, output np.ndarray with np.float32"""
-        if 'BGE' in str(type(self.client)):
+        if 'bge' in self._type:
+            import torch
             with torch.no_grad():
                 feature = self.client.encode(text=text, image=path)
                 return feature.cpu().numpy().astype(np.float32)
-        else:
+        elif 'bce' in self._type:
             if text is None:
                 raise ValueError('This model only support text')
             emb = self.client.encode([text], show_progress_bar=False, normalize_embeddings=True)
@@ -57,3 +70,8 @@ class Embedder:
             # for norm in np.linalg.norm(emb, axis=1):
             #     assert abs(norm - 1) < 0.001
             return emb
+        else:
+            # siliconcloud bce API
+            if text is None:
+                raise ValueError('This model only support text')
+            # TODO
